+++ driven/storage/adapter_example.go
@@ -71,3 +71,22 @@ func (a *Adapter) DeleteExample(orgID string, appID string, id string) error {
 
 	return nil
 }
+
+// InsertExample inserts a new example
+func (a *Adapter) SaveLegacyEvents(legacyEvents []model.LegacyEvent) error {
+	records := []interface{}{}
+	for _, event := range legacyEvents {
+		records = append(records, legacyEventToStorage(event))
+		//records := legacyEventsToStorage(legacyEvents)
+
+		//_, err := a.db.legacyEvents.InsertMany()
+		if len(records) == 100 {
+			_, err := a.db.legacyEvents.InsertMany(nil, records, nil)
+			if err != nil {
+				return err
+			}
+			//		records = []interface{}{}
+		}
+	}
+	return nil
+}